requirements.txt
amqp==2.6.1
billiard==3.6.3.0
blinker==1.4
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Cors==1.10.3
Flask-Mail==0.9.1
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.4
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
kombu==4.6.11
MarkupSafe==1.1.1
mysqlclient==2.0.1
numpy==1.19.1
openpyxl==3.0.4
pandas==1.1.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
urllib3==1.25.10
vine==1.3.0
Werkzeug==1.0.1
zipp==3.1.0
